requirements.txt
astroid==2.3.2
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.6
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
Pygments==2.4.2
pylint==2.4.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.4
wrapt==1.11.2
